03_Mistral_bot_nomem/requirements.txt
langchain
langchain-community
langchain-core
mistralai
python-dotenv
fastapi
uvicorn
pydantic
requests